+++ extension-recorder/src/main/java/io/github/jwdeveloper/tiktok/extension/recorder/api/LiveRecorder.java
@@ -26,6 +26,5 @@ import io.github.jwdeveloper.tiktok.listener.TikTokEventListener;
 
 public interface LiveRecorder extends TikTokEventListener {
 
-    void pause();
-    void unpause();
+
 }
